+++ examples/mongoengine/main.py
@@ -178,12 +178,16 @@ def create_example_data():
     ).save()
 
 
-with MongoDbContainer("mongo:7.0.7") as mongo:
-    mongo_uri = mongo.get_connection_url()
-    connect(host=mongo_uri)
-    create_example_data()
-    admin.add_view(UserView(User, "User"))
-    admin.add_view(TweetView(Tweet, "Tweets"))
+admin.add_view(UserView(User, "User"))
+admin.add_view(TweetView(Tweet, "Tweets"))
+
 
 if __name__ == "__main__":
-    app.run(debug=True)
+    with MongoDbContainer("mongo:7.0.7") as mongo:
+        mongo_uri = mongo.get_connection_url()
+        print("using mongo uri:", mongo_uri)
+
+        connect(host=mongo_uri)
+        create_example_data()
+
+        app.run(debug=True)
