+++ examples/sqla/admin/main.py
@@ -1,5 +1,3 @@
-from flask import redirect
-from flask import url_for
 from flask_admin import Admin
 from flask_admin.babel import gettext
 from flask_admin.base import MenuLink
@@ -19,31 +17,6 @@ from .models import Post
 from .models import Tag
 from .models import Tree
 from .models import User
-
-
-@app.route("/")
-def index():
-    tmp = """
-<p><a href="/admin/?lang=en">Click me to get to Admin! (English)</a></p>
-<p><a href="/admin/?lang=cs">Click me to get to Admin! (Czech)</a></p>
-<p><a href="/admin/?lang=de">Click me to get to Admin! (German)</a></p>
-<p><a href="/admin/?lang=es">Click me to get to Admin! (Spanish)</a></p>
-<p><a href="/admin/?lang=fa">Click me to get to Admin! (Farsi)</a></p>
-<p><a href="/admin/?lang=fr">Click me to get to Admin! (French)</a></p>
-<p><a href="/admin/?lang=pt">Click me to get to Admin! (Portuguese)</a></p>
-<p><a href="/admin/?lang=ru">Click me to get to Admin! (Russian)</a></p>
-<p><a href="/admin/?lang=pa">Click me to get to Admin! (Punjabi)</a></p>
-<p><a href="/admin/?lang=zh_CN">Click me to get to Admin! (Chinese - Simplified)</a></p>
-<p>
-  <a href="/admin/?lang=zh_TW">Click me to get to Admin! (Chinese - Traditional)</a>
-</p>
-"""
-    return tmp
-
-
-@app.route("/favicon.ico")
-def favicon():
-    return redirect(url_for("static", filename="/favicon.ico"))
 
 
 # Custom filter class
